requirements.txt
anthropic==0.94.0
beautifulsoup4==4.14.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
openpyxl==3.1.5
pandas==2.3.3
python-dotenv==1.2.2
python-jobspy==1.1.82
requests==2.33.1
